move impoer manager into individual file.
